Rename to follow the project's filename pattern
